+++ example/lib/main.dart
@@ -18,6 +18,7 @@ class ExampleArcaneApp extends StatelessWidget {
               builder: (context) => PrimaryButton(
                   child: Text("Date Range Picker"),
                   onPressed: () => DialogCommand(
+                        hint: "Select a date range",
                         leading: Icon(Icons.airplane),
                         trailing: Icon(Icons.airplane),
                         onConfirm: (range) {