+++ lib/component/screen/navigation_screen.dart
@@ -14,7 +14,7 @@ class NavTab {
   });
 }
 
-enum NavigationType { bottomNavigationBar, navigationRail, sidebar }
+enum NavigationType { bottomNavigationBar, navigationRail, sidebar, drawer }
 
 class NavigationScreen extends AbstractStatelessScreen {
   final int index;
@@ -24,6 +24,7 @@ class NavigationScreen extends AbstractStatelessScreen {
   final List<NavTab> tabs;
   final Widget? overrideSidebarGap;
   final double siderailTopPadding;
+  final bool endSide;
   const NavigationScreen(
       {super.key,
       this.siderailRightPadding = 8,
@@ -32,6 +33,7 @@ class NavigationScreen extends AbstractStatelessScreen {
       this.siderailTopPadding = 8,
       this.onIndexChanged,
       required this.tabs,
+      this.endSide = false,
       this.type = NavigationType.bottomNavigationBar});
 
   Widget buildBottomNavigationBar(BuildContext context, int index) => ButtonBar(
@@ -77,7 +79,8 @@ class NavigationScreen extends AbstractStatelessScreen {
         ],
       ).padTop(siderailTopPadding);
 
-  Widget buildSidebar(BuildContext context, int index) => NavigationSidebar(
+  Widget buildSidebar(BuildContext context, int index, {bool drawer = false}) =>
+      NavigationSidebar(
         labelType: NavigationLabelType.expanded,
         constraints: const BoxConstraints(minWidth: 100, maxWidth: 150),
         index: index,
@@ -93,6 +96,10 @@ class NavigationScreen extends AbstractStatelessScreen {
           ],
           ...tabs.mapIndexed((e, i) => NavigationButton(
               onChanged: (e) {
+                if (drawer) {
+                  Arcane.closeDrawer(context);
+                }
+
                 if (index != i) {
                   onIndexChanged?.call(i);
                 }
@@ -104,43 +111,77 @@ class NavigationScreen extends AbstractStatelessScreen {
       ).padTop(siderailTopPadding);
 
   @override
-  Widget build(BuildContext context) => IndexedStack(
-        index: index,
-        children: tabs
-            .mapIndexed((tab, index) => switch (type) {
-                  NavigationType.bottomNavigationBar => tabs[index].builder(
-                      context, buildBottomNavigationBar(context, index)),
-                  NavigationType.sidebar => Scaffold(
-                      child: Row(
-                        mainAxisSize: MainAxisSize.max,
-                        crossAxisAlignment: CrossAxisAlignment.stretch,
-                        children: [
-                          buildSidebar(context, index),
-                          overrideSidebarGap ?? Gap(siderailRightPadding),
-                          Expanded(
-                            child: BlockBackButton(
-                                builder: (context) =>
-                                    tabs[index].builder(context, null)),
-                          ),
+  Widget build(BuildContext context) => DrawerOverlay(
+          child: Pylon<NavigationType>(
+        value: type,
+        local: true,
+        builder: (context) => IndexedStack(
+          index: index,
+          children: tabs
+              .mapIndexed((tab, index) => switch (type) {
+                    NavigationType.drawer => InjectBarEnds(
+                        trailing: endSide,
+                        start: !endSide,
+                        children: (context) => [
+                          IconButton(
+                              icon: Icon(Icons.menu_ionic),
+                              onPressed: () {
+                                openDrawer(
+                                    expands: false,
+                                    showDragHandle: false,
+                                    context: context,
+                                    builder: (context) => Column(
+                                          mainAxisSize: MainAxisSize.max,
+                                          children: [
+                                            SingleChildScrollView(
+                                              child: buildSidebar(
+                                                  context, index,
+                                                  drawer: true),
+                                            )
+                                          ],
+                                        ),
+                                    position: endSide
+                                        ? OverlayPosition.right
+                                        : OverlayPosition.left);
+                              })
                         ],
+                        builder: (context) =>
+                            tabs[index].builder(context, null),
                       ),
-                    ),
-                  NavigationType.navigationRail => Scaffold(
-                      child: Row(
-                        mainAxisSize: MainAxisSize.max,
-                        crossAxisAlignment: CrossAxisAlignment.stretch,
-                        children: [
-                          buildNavigationRail(context, index),
-                          Gap(siderailRightPadding),
-                          Expanded(
-                            child: BlockBackButton(
-                                builder: (context) =>
-                                    tabs[index].builder(context, null)),
-                          ),
-                        ],
+                    NavigationType.bottomNavigationBar => tabs[index].builder(
+                        context, buildBottomNavigationBar(context, index)),
+                    NavigationType.sidebar => Scaffold(
+                        child: Row(
+                          mainAxisSize: MainAxisSize.max,
+                          crossAxisAlignment: CrossAxisAlignment.stretch,
+                          children: [
+                            buildSidebar(context, index),
+                            overrideSidebarGap ?? Gap(siderailRightPadding),
+                            Expanded(
+                              child: BlockBackButton(
+                                  builder: (context) =>
+                                      tabs[index].builder(context, null)),
+                            ),
+                          ],
+                        ),
                       ),
-                    ),
-                })
-            .toList(),
-      );
+                    NavigationType.navigationRail => Scaffold(
+                        child: Row(
+                          mainAxisSize: MainAxisSize.max,
+                          crossAxisAlignment: CrossAxisAlignment.stretch,
+                          children: [
+                            buildNavigationRail(context, index),
+                            Gap(siderailRightPadding),
+                            Expanded(
+                              child: BlockBackButton(
+                                  builder: (context) =>
+                                      tabs[index].builder(context, null)),
+                            ),
+                          ],
+                        ),
+                      ),
+                  })
+              .toList(),
+        ),
+      ));
 }